# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Clinic Platform API
  version: 0.1.0
  summary: Phase 00 foundation slice — health, readiness, version, and a gated diagnostics path.
  description: |
    This document is the source of truth for the HTTP contract (ADR 0003).
    Server and client code both conform to it; neither generates it.

    Phase 01 adds authentication, identity projection, and deny-by-default
    capabilities. Clinical, pharmacy, and AI capabilities remain absent.

    Compatibility rules (ADR 0003):
      * Additive optional fields are compatible within a version.
      * A removal, a type change, a narrowed enum, or a new required field is
        breaking and requires a new version plus a dual-read migration.
      * Removals require deprecation telemetry proving no live consumer.
  contact:
    name: Platform architecture
    url: https://example.invalid/clinic/platform
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
  - url: https://staging.clinic.invalid
    description: Staging (synthetic data only)
...
paths:
...
  /api/v1/auth/otp-requests:
    post:
      tags: [auth]
      operationId: requestOtp
      summary: Request an OTP for a declared purpose
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '200':
          description: Generic OTP required envelope.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OtpChallengeResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/otp-verifications:
    post:
      tags: [auth]
      operationId: verifyOtp
      summary: Consume a one-time code
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerificationRequest'
      responses:
        '200':
          description: Restricted session, recovery continuation, or generic OTP envelope.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthSessionResult'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/auth/mfa/totp/enroll:
    post:
      tags: [auth]
      operationId: enrollTotp
      summary: Start self-service TOTP enrollment
      security:
        - deviceToken: []
        - adminSession: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Provisioning URI once. Never logged.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MfaEnrollResult'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/mfa/totp/confirm:
    post:
      tags: [auth]
      operationId: confirmTotp
      summary: Confirm TOTP enrollment and issue recovery codes once
      security:
        - deviceToken: []
        - adminSession: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: Recovery codes shown once.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MfaConfirmResult'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/auth/mfa/totp/disable:
    post:
      tags: [auth]
      operationId: disableTotp
      summary: Disable the active TOTP factor
      security:
        - deviceToken: []
        - adminSession: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthAck'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque device access token issued after OTP or password+MFA for
        patient_mobile, doctor_desktop, and pharmacy_desktop. Stored hashed
        server-side. Never used by the admin browser.
    adminSession:
      type: apiKey
      in: cookie
      name: clinic_session
      description: |
        HTTP-only, Secure, SameSite session cookie for the React admin, paired
        with a CSRF token header. The admin never stores a bearer token in local
        storage. Device clients must not send this cookie as their credential.
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Client-supplied correlation identifier. When absent the server assigns
        one. Always echoed in the response body and the `X-Request-Id` header.
      schema:
        $ref: '#/components/schemas/Uuid'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Cryptographically random key generated per user intent and reused only
        for retries of the identical request. Scoped server-side to the
        authenticated actor/device, the operation, and the tenant where
        applicable.
      schema:
        type: string
        minLength: 16
        maxLength: 255
        pattern: '^[A-Za-z0-9._~-]+$'
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: Language negotiation. Supported tags are `ar` and `en`.
      schema:
        type: string
        default: en
        examples: ['ar', 'en', 'ar-EG', 'en-US']
...
  headers:
...
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        minimum: 0
...
  schemas:
...
    Uuid:
      type: string
      format: uuid
      description: UUIDv7 rendered as a string. Public APIs never expose sequential identifiers (ADR 0005).
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      examples: ['0199a5c8-1f2e-7c3a-9b41-2f6d0c5e7a10']
    Instant:
      type: string
      format: date-time
      description: |
        RFC 3339 / ISO-8601 instant with an explicit offset, always stored and
        transmitted in UTC. Never an ambiguous local date-time.
      examples: ['2026-08-24T19:05:00Z']
...
    Envelope:
      type: object
      description: |
        Every response uses this envelope (`plan.md` section 106). `data` carries
        the payload, `meta` carries pagination and negotiation information,
        `errors` is empty on success, and `request_id` correlates the request
        across logs, traces, and support conversations.
      required: [data, meta, errors, request_id]
      properties:
        data:
          description: Payload. `null` for operations with no return value.
        meta:
          $ref: '#/components/schemas/Meta'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
          default: []
        request_id:
          $ref: '#/components/schemas/Uuid'
    ErrorEnvelope:
      type: object
      description: Failure envelope. `data` is null and `errors` has at least one entry.
      required: [data, meta, errors, request_id]
      properties:
        data:
          type: 'null'
        meta:
          $ref: '#/components/schemas/Meta'
        errors:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Error'
        request_id:
          $ref: '#/components/schemas/Uuid'
    Meta:
      type: object
      description: Response metadata. Never carries business payload.
      additionalProperties: true
      properties:
        locale:
          type: string
          enum: [ar, en]
        pagination:
          $ref: '#/components/schemas/CursorPagination'
        deprecation:
          $ref: '#/components/schemas/Deprecation'
    CursorPagination:
      type: object
      description: Cursor pagination for large or mutable collections. Offset pagination is not used.
      required: [has_more]
      additionalProperties: false
      properties:
        has_more:
          type: boolean
        next:
          type: [string, 'null']
          maxLength: 512
          description: Opaque cursor for the next page, or null when `has_more` is false.
        limit:
          type: integer
          minimum: 1
          maximum: 100
    Deprecation:
      type: object
      description: Present when the caller used a deprecated field or operation. Drives deprecation telemetry (ADR 0003).
      required: [message]
      properties:
        message:
          type: string
          maxLength: 500
        sunset_at:
          $ref: '#/components/schemas/Instant'
    Error:
      type: object
      description: |
        A single failure. Errors expose a stable machine code and a safe human
        message. They never expose a stack trace, SQL, an object key, a provider
        payload, or the existence of a protected resource.
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          maxLength: 500
          description: Safe, localized, human-readable message. Never contains sensitive values.
        field:
          type: [string, 'null']
          maxLength: 200
          description: JSON pointer or dotted path for a field-level validation failure.
        meta:
          type: object
          description: Bounded, non-sensitive detail such as an allowed range.
          additionalProperties: true
    ErrorCode:
      type: string
      description: |
        Stable machine code. Clients branch on this, never on the message.

        Status mapping (phase file "API contract"):
          400 malformed protocol or input
          401 unauthenticated
          403 denied or CSRF mismatch
          404 hiding resource existence is safer
          409 state, version, or idempotency conflict
          422 field or business validation
          429 throttled
          5xx server or dependency failure
      enum:
        - MALFORMED_REQUEST
        - UNSUPPORTED_MEDIA_TYPE
        - REQUEST_TOO_LARGE
        - UNAUTHENTICATED
        - CSRF_MISMATCH
        - TOKEN_EXPIRED
        - PERMISSION_DENIED
        - NOT_FOUND
        - STATE_CONFLICT
        - VERSION_CONFLICT
        - IDEMPOTENCY_KEY_REUSED
        - IDEMPOTENCY_IN_PROGRESS
        - VALIDATION_FAILED
        - CURSOR_INVALID
        - UNSUPPORTED_SCHEMA_VERSION
        - RATE_LIMITED
        - DEPENDENCY_UNAVAILABLE
        - INTERNAL_ERROR

    # ------------------------------------------------------------ operational
...
    OtpRequest:
      type: object
      required: [phone, purpose]
      additionalProperties: false
      properties:
        phone:
          type: string
          maxLength: 32
        purpose:
          type: string
          enum: [registration, phone_change, recovery, profile_claim]
        language:
          type: string
          enum: [ar, en]
    OtpVerificationRequest:
      type: object
      required: [challenge_id, code, client_class, platform, device_label]
      additionalProperties: false
      properties:
        challenge_id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
          pattern: '^[0-9]{6}$'
        client_class:
          $ref: '#/components/schemas/ClientClass'
        platform:
          $ref: '#/components/schemas/DevicePlatform'
        device_label:
          type: string
          maxLength: 120
...
    MfaVerifyRequest:
      type: object
      required: [code]
      additionalProperties: false
      properties:
        code:
          type: string
          pattern: '^[0-9]{6}$'
    MfaEnrollResult:
      type: object
      required: [factor_id, provisioning_uri]
      additionalProperties: false
      properties:
        factor_id:
          $ref: '#/components/schemas/Uuid'
        provisioning_uri:
          type: string
          maxLength: 512
    MfaConfirmResult:
      type: object
      required: [verified, recovery_codes]
      additionalProperties: false
      properties:
        verified:
          type: boolean
        recovery_codes:
          type: array
          maxItems: 16
          items:
            type: string
            maxLength: 32
...
    ClientClass:
      type: string
      enum: [patient_mobile, doctor_desktop, pharmacy_desktop, admin_web]
    DevicePlatform:
      type: string
      enum: [android, ios, windows, macos, linux, web]
    OtpChallengeResult:
      type: object
      required: [challenge_id, status]
      additionalProperties: false
      properties:
        challenge_id:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum: [otp_required, recovery_verified]
    AuthSessionResult:
      type: object
      additionalProperties: false
      properties:
        mfa_required:
          type: boolean
        challenge_id:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          maxLength: 32
        session_kind:
          type: string
          enum: [device, admin_cookie]
        session_id:
          $ref: '#/components/schemas/Uuid'
        device_id:
          $ref: '#/components/schemas/Uuid'
        user_id:
          $ref: '#/components/schemas/Uuid'
        access_token:
          type: string
          maxLength: 256
          description: Present once for device sessions. Never logged.
        refresh_token:
          type: string
          maxLength: 256
        expires_in:
          type: integer
          minimum: 1
        capabilities:
          type: array
          items:
            type: string
            maxLength: 120
        assurance_level:
          type: string
          maxLength: 32
...
    AuthAck:
      type: object
      additionalProperties: false
      properties:
        revoked:
          type: boolean
        credential_rotated:
          type: boolean
        status:
          type: string
          enum: [applied, cooling_off, manual_review, completed]
          maxLength: 32
        csrf:
          type: boolean
        disabled:
          type: boolean
        recovery_codes:
          type: array
          maxItems: 16
          items:
            type: string
            maxLength: 32
  responses:
...
    Unauthenticated:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Not found, or hidden because disclosing existence is unsafe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UnprocessableEntity:
      description: Field-level or business validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            validation:
              value:
                data: null
                meta: { locale: en }
                errors:
                  - code: VALIDATION_FAILED
                    message: The label may contain only letters, digits, spaces, underscores, and hyphens.
                    field: label
                request_id: 0199a5c8-1f2e-7c3a-9b41-2f6d0c5e7a10
    TooManyRequests:
      description: Throttled.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    AuthAck:
      description: Bounded acknowledgement with no secrets.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthAck'
security: []